# application/services/custom_services/apply/src/models/entitymodels/name.ts
export interface name 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: String,
   email: String
}